Reorganize auton packages
