+++ src/main/java/frc/team4373/robot/Robot.java
@@ -8,7 +8,7 @@ import edu.wpi.first.wpilibj.smartdashboard.SendableChooser;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import frc.team4373.robot.RobotMap.CargoShipPort;
 import frc.team4373.robot.RobotMap.Side;
-import frc.team4373.robot.commands.auton.elemental.RetractClimberAuton;
+import frc.team4373.robot.commands.auton.climb.RetractClimberAuton;
 import frc.team4373.robot.commands.auton.sequences.*;
 import frc.team4373.robot.subsystems.*;
