+++ src/main/java/frc/team4373/robot/commands/auton/climb/DeployClimberFrontAuton.java
@@ -1,4 +1,4 @@
-package frc.team4373.robot.commands.auton.elemental;
+package frc.team4373.robot.commands.auton.climb;
 
 import com.ctre.phoenix.motorcontrol.NeutralMode;
 import edu.wpi.first.wpilibj.command.Command;
@@ -7,11 +7,9 @@ import frc.team4373.robot.subsystems.ClimberDrive;
 import frc.team4373.robot.subsystems.Drivetrain;
 
 /**
- * Retracts the climber and sets the neutral mode of the drivetrain to brake.
- *
- * <p>Should only be used for testing.
+ * Extends the font climber and sets neutral modes.
  */
-public class RetractClimberAuton extends Command {
+public class DeployClimberFrontAuton extends Command {
     private Climber climber;
     private ClimberDrive cld;
     private Drivetrain drivetrain;
@@ -19,7 +17,7 @@ public class RetractClimberAuton extends Command {
     /**
      * Constructs a command to retract the climber.
      */
-    public RetractClimberAuton() {
+    public DeployClimberFrontAuton() {
         requires(this.climber = Climber.getInstance());
         requires(this.drivetrain = Drivetrain.getInstance());
         requires(this.cld = ClimberDrive.getInstance());
@@ -32,9 +30,9 @@ public class RetractClimberAuton extends Command {
 
     @Override
     protected void execute() {
-        this.climber.retractAll();
-        this.drivetrain.setNeutralMode(NeutralMode.Brake);
-        this.cld.setNeutralMode(NeutralMode.Coast);
+        this.climber.deployFront();
+        this.drivetrain.setNeutralMode(NeutralMode.Coast);
+        this.cld.setNeutralMode(NeutralMode.Brake);
     }
 
     @Override
