+++ src/main/java/frc/team4373/robot/commands/auton/climb/RetractClimberRearAuton.java
@@ -1,4 +1,4 @@
-package frc.team4373.robot.commands.auton.elemental;
+package frc.team4373.robot.commands.auton.climb;
 
 import com.ctre.phoenix.motorcontrol.NeutralMode;
 import edu.wpi.first.wpilibj.command.Command;
@@ -7,20 +7,20 @@ import frc.team4373.robot.subsystems.ClimberDrive;
 import frc.team4373.robot.subsystems.Drivetrain;
 
 /**
- * Retracts the front half of the climber and sets the neutral mode of the drivetrain to coast.
+ * Retracts the rear half of the climber and sets the neutral mode of the drivetrain to brake.
  */
-public class RetractClimberFrontAuton extends Command {
+public class RetractClimberRearAuton extends Command {
     private Climber climber;
     private ClimberDrive cld;
     private Drivetrain drivetrain;
 
     /**
-     * Constructs a command to retract the front half of the climber.
+     * Constructs a command to retract the rear half of the climber.
      */
-    public RetractClimberFrontAuton() {
+    public RetractClimberRearAuton() {
         requires(this.climber = Climber.getInstance());
-        requires(this.drivetrain = Drivetrain.getInstance());
         requires(this.cld = ClimberDrive.getInstance());
+        requires(this.drivetrain = Drivetrain.getInstance());
     }
 
     @Override
@@ -30,7 +30,7 @@ public class RetractClimberFrontAuton extends Command {
 
     @Override
     protected void execute() {
-        this.climber.retractFront();
+        this.climber.retractRear();
         this.drivetrain.setNeutralMode(NeutralMode.Brake);
         this.cld.setNeutralMode(NeutralMode.Coast);
     }
