+++ src/main/java/frc/team4373/robot/commands/auton/intake/CollectCargoAuton.java
@@ -1,40 +1,28 @@
-package frc.team4373.robot.commands.auton.elemental;
+package frc.team4373.robot.commands.auton.intake;
 
 import edu.wpi.first.wpilibj.command.Command;
 import frc.team4373.robot.subsystems.Intake;
 
-/**
- * Deploys the intake.
- */
-public class DeployIntakeAuton extends Command {
+public class CollectCargoAuton extends Command {
+
     private Intake intake;
 
-    public DeployIntakeAuton() {
+    public CollectCargoAuton() {
         requires(this.intake = Intake.getInstance());
     }
 
     @Override
     protected void initialize() {
-        setTimeout(0.25);
+        setTimeout(0.5);
     }
 
     @Override
     protected void execute() {
-        this.intake.deploy();
+        this.intake.collectCargo();
     }
 
     @Override
     protected boolean isFinished() {
         return this.isTimedOut();
-    }
-
-    @Override
-    protected void end() {
-
-    }
-
-    @Override
-    protected void interrupted() {
-        this.end();
     }
 }
