+++ src/main/java/frc/team4373/robot/commands/auton/sequences/DriveForwardAuton.java
@@ -3,7 +3,7 @@ package frc.team4373.robot.commands.auton.sequences;
 import edu.wpi.first.wpilibj.command.CommandGroup;
 import frc.team4373.robot.RobotMap;
 import frc.team4373.robot.commands.auton.drive.DriveDistanceAuton;
-import frc.team4373.robot.commands.auton.elemental.DeployIntakeAuton;
+import frc.team4373.robot.commands.auton.intake.DeployIntakeAuton;
 
 public class DriveForwardAuton extends CommandGroup {
     public DriveForwardAuton() {
