+++ src/main/java/frc/team4373/robot/commands/auton/sequences/RHatchAuton.java
@@ -6,10 +6,10 @@ import frc.team4373.robot.commands.auton.drive.ApproachVisionTargetAuton;
 import frc.team4373.robot.commands.auton.drive.DriveDistanceAuton;
 import frc.team4373.robot.commands.auton.drive.MiddleWheelAdjusterAuton;
 import frc.team4373.robot.commands.auton.drive.TurnToAngleAuton;
-import frc.team4373.robot.commands.auton.elemental.CollectHatchPanelAuton;
-import frc.team4373.robot.commands.auton.elemental.DeployIntakeAuton;
-import frc.team4373.robot.commands.auton.elemental.ReleaseHatchPanelAuton;
-import frc.team4373.robot.commands.auton.elemental.SetLiftAuton;
+import frc.team4373.robot.commands.auton.intake.CollectHatchPanelAuton;
+import frc.team4373.robot.commands.auton.intake.DeployIntakeAuton;
+import frc.team4373.robot.commands.auton.intake.ReleaseHatchPanelAuton;
+import frc.team4373.robot.commands.auton.lift.SetLiftAuton;
 
 public class RHatchAuton extends CommandGroup {
     private RobotMap.Side side;
